requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.10.0
django-extensions==3.1.5
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
